add holo

+++ rust/autd3-core/src/stm/mod.rs
@@ -4,7 +4,7 @@
  * Created Date: 05/05/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 07/03/2023
+ * Last Modified: 11/05/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -14,25 +14,25 @@
 mod focus;
 mod gain;
 
-use autd3_driver::FPGA_CLK_FREQ;
+use autd3_driver::{float, FPGA_CLK_FREQ};
 pub use focus::FocusSTM;
 pub use gain::GainSTM;
 
 pub trait STM {
     fn size(&self) -> usize;
-    fn set_freq(&mut self, freq: f64) -> f64 {
-        let sample_freq = self.size() as f64 * freq;
-        self.set_sampling_freq_div((FPGA_CLK_FREQ as f64 / sample_freq) as _);
+    fn set_freq(&mut self, freq: float) -> float {
+        let sample_freq = self.size() as float * freq;
+        self.set_sampling_freq_div((FPGA_CLK_FREQ as float / sample_freq) as _);
         STM::freq(self)
     }
-    fn freq(&self) -> f64 {
-        self.sampling_freq() / self.size() as f64
+    fn freq(&self) -> float {
+        self.sampling_freq() / self.size() as float
     }
-    fn sampling_freq(&self) -> f64 {
-        FPGA_CLK_FREQ as f64 / self.sampling_freq_div() as f64
+    fn sampling_freq(&self) -> float {
+        FPGA_CLK_FREQ as float / self.sampling_freq_div() as float
     }
-    fn set_sampling_freq(&mut self, sample_freq: f64) -> f64 {
-        self.set_sampling_freq_div((FPGA_CLK_FREQ as f64 / sample_freq) as _);
+    fn set_sampling_freq(&mut self, sample_freq: float) -> float {
+        self.set_sampling_freq_div((FPGA_CLK_FREQ as float / sample_freq) as _);
         STM::sampling_freq(self)
     }
     fn set_sampling_freq_div(&mut self, freq_div: u32);
